api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Convrt API service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.119.0",
    "uvicorn==0.37.0",
    "mangum==0.19.0",
    "python-multipart==0.0.20",
    "pillow==12.0.0",
    "pymongo==4.15.3",
    "bson==0.5.10",
    "email-validator==2.3.0",
    "passlib==1.7.4",
    "bson==0.5.10",
    "bcrypt==4.2.1",
    "PyJWT==2.10.0",
]
